Removed sf::Vector2f origin from QuadTree class, changed constructors accordingly. Changed how the QuadTree subnodes were being constructed in QuadTree::split() to work with removal of origin vector.
We can use the bounding rectangle shape that we use to draw the QuadTree instead of an origin vector. When calling sf::RectangleShape::getPosition() it returns a vector at the origin of the shape. The origin of the shape is default initialized at the top left or the (0,0) coordinate relative to the shape itself.