Added sf::Color member to Particle class and removed sf::VertexArray with duplicate data
Before I was using a vertex array for each particle to store the color,
and writing the particle position to each vertex as well as the color.
This meant that we allocated data for the position vector twice: once in
the particle class and once for the vertex in the vertex array associated
with the particle.

I just added a color member to the particle itself allowing for the removal
or the vertex array entirely.