Added new struct that represents a single particle within a quadtree node
This struct will be used as the type in a container like std::vector,
and contains a integer next_particle that is the next particle in the current leaf node,
and particle_index which is a handle to the actual particle element in the particles vector
for the ParticleSimulation class.

This struct will allow us to get rid of the std::vector<particle*> in each quadtree node that is
used to reference each particle within the leaf node. Instead we can keep an integer handle in the Node struct
to the first ParticleElementNode within the leaf, and the next_particle member within the ParticleElementNode
allows us to use it as a singly-linked list to all other particles within the leaf.