Major fixes and overhaul to QuadTree class
Major fixes to the init function that sets up the tree.

All function also traverse the tree with the help of a Node struct that has the
depth, position, and size of the current node being processed. Since we can track
these parameters with the struct each time a traversal is needed, we can remove these
members from the QuadTree::Node struct that is held in the array.

The insert and split function was also updated so they do not call each other and splitting
the nodes is done more gracefully.

Ultimately this commit yeilded a great speedup.