QuadTree class now uses FreeList to keep track of gravity COM and mass
We can take the gravity and total mass members out of the base Node struct
and instead create a new struct with these members. We can store this
struct in a FreeList for every leaf node, removing the appropriate node
on a split and adding the childrens nodes. This means we effectively reduce the
memory used for storing gravity COM and mass from O(n) where n is the total amount
of nodes in the quadtree array, to O(pow(4,maxTreeDepth)).